Update from original

+++ modules/siddhi-core/src/main/java/org/wso2/siddhi/core/query/selector/attribute/aggregator/StddevAttributeAggregator.java
@@ -1,17 +1,19 @@
 /*
  * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
  * You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 
 package org.wso2.siddhi.core.query.selector.attribute.aggregator;
@@ -60,10 +62,14 @@ public class StddevAttributeAggregator extends AttributeAggregator {
     }
 
     @Override
-    public Attribute.Type getReturnType() { return stddevOutputAttributeAggregator.getReturnType(); }
+    public Attribute.Type getReturnType() {
+        return stddevOutputAttributeAggregator.getReturnType();
+    }
 
     @Override
-    public Object processAdd(Object data) { return stddevOutputAttributeAggregator.processAdd(data); }
+    public Object processAdd(Object data) {
+        return stddevOutputAttributeAggregator.processAdd(data);
+    }
 
     @Override
     public Object processAdd(Object[] data) {
@@ -71,7 +77,9 @@ public class StddevAttributeAggregator extends AttributeAggregator {
     }
 
     @Override
-    public Object processRemove(Object data) { return stddevOutputAttributeAggregator.processRemove(data); }
+    public Object processRemove(Object data) {
+        return stddevOutputAttributeAggregator.processRemove(data);
+    }
 
     @Override
     public Object processRemove(Object[] data) {
@@ -79,19 +87,27 @@ public class StddevAttributeAggregator extends AttributeAggregator {
     }
 
     @Override
-    public Object reset() { return stddevOutputAttributeAggregator.reset(); }
+    public Object reset() {
+        return stddevOutputAttributeAggregator.reset();
+    }
 
     @Override
-    public Object[] currentState() { return stddevOutputAttributeAggregator.currentState(); }
+    public Object[] currentState() {
+        return stddevOutputAttributeAggregator.currentState();
+    }
 
     @Override
-    public void start() {}
+    public void start() {
+    }
 
     @Override
-    public void stop() {}
+    public void stop() {
+    }
 
     @Override
-    public void restoreState(Object[] state) { stddevOutputAttributeAggregator.restoreState(state); }
+    public void restoreState(Object[] state) {
+        stddevOutputAttributeAggregator.restoreState(state);
+    }
 
     private class StddevAttributeAggregatorDouble extends StddevAttributeAggregator {
         private final Attribute.Type type = Attribute.Type.DOUBLE;
@@ -99,7 +115,9 @@ public class StddevAttributeAggregator extends AttributeAggregator {
         private int count = 0;
 
         @Override
-        public Attribute.Type getReturnType() { return type; }
+        public Attribute.Type getReturnType() {
+            return type;
+        }
 
         @Override
         public Object processAdd(Object data) {
@@ -114,7 +132,7 @@ public class StddevAttributeAggregator extends AttributeAggregator {
                 oldMean = mean;
                 sum += value;
                 mean = sum / count;
-                stdDeviation += (value - oldMean)*(value - mean);
+                stdDeviation += (value - oldMean) * (value - mean);
             }
 
             if (count < 2) {
@@ -129,13 +147,13 @@ public class StddevAttributeAggregator extends AttributeAggregator {
             double value = (Double) data;
 
             if (count == 0) {
-                sum = mean = 0;
-                stdDeviation = 0;
+                sum = mean = 0.0;
+                stdDeviation = 0.0;
             } else {
                 oldMean = mean;
                 sum -= value;
                 mean = sum / count;
-                stdDeviation -= (value - oldMean)*(value - mean);
+                stdDeviation -= (value - oldMean) * (value - mean);
             }
 
             if (count < 2) {
@@ -149,11 +167,13 @@ public class StddevAttributeAggregator extends AttributeAggregator {
             sum = mean = oldMean = 0.0;
             stdDeviation = 0.0;
             count = 0;
-            return 0;
+            return 0.0;
         }
 
         @Override
-        public Object[] currentState() { return new Object[] {sum, mean, oldMean, stdDeviation, count}; }
+        public Object[] currentState() {
+            return new Object[]{sum, mean, oldMean, stdDeviation, count};
+        }
 
         @Override
         public void restoreState(Object[] state) {
@@ -166,12 +186,14 @@ public class StddevAttributeAggregator extends AttributeAggregator {
     }
 
     private class StddevAttributeAggregatorFloat extends StddevAttributeAggregator {
-        private final Attribute.Type type = Attribute.Type.FLOAT;
+        private final Attribute.Type type = Attribute.Type.DOUBLE;
         private double mean, oldMean, stdDeviation, sum;
         private int count = 0;
 
         @Override
-        public Attribute.Type getReturnType() { return type; }
+        public Attribute.Type getReturnType() {
+            return type;
+        }
 
         @Override
         public Object processAdd(Object data) {
@@ -186,7 +208,7 @@ public class StddevAttributeAggregator extends AttributeAggregator {
                 oldMean = mean;
                 sum += value;
                 mean = sum / count;
-                stdDeviation += (value - oldMean)*(value - mean);
+                stdDeviation += (value - oldMean) * (value - mean);
             }
 
             if (count < 2) {
@@ -201,13 +223,13 @@ public class StddevAttributeAggregator extends AttributeAggregator {
             double value = (Float) data;
 
             if (count == 0) {
-                sum = mean = 0;
-                stdDeviation = 0;
+                sum = mean = 0.0;
+                stdDeviation = 0.0;
             } else {
                 oldMean = mean;
                 sum -= value;
                 mean = sum / count;
-                stdDeviation -= (value - oldMean)*(value - mean);
+                stdDeviation -= (value - oldMean) * (value - mean);
             }
 
             if (count < 2) {
@@ -221,11 +243,13 @@ public class StddevAttributeAggregator extends AttributeAggregator {
             sum = mean = oldMean = 0.0;
             stdDeviation = 0.0;
             count = 0;
-            return 0;
+            return 0.0;
         }
 
         @Override
-        public Object[] currentState() { return new Object[] {sum, mean, oldMean, stdDeviation, count}; }
+        public Object[] currentState() {
+            return new Object[]{sum, mean, oldMean, stdDeviation, count};
+        }
 
         @Override
         public void restoreState(Object[] state) {
@@ -238,12 +262,14 @@ public class StddevAttributeAggregator extends AttributeAggregator {
     }
 
     private class StddevAttributeAggregatorInt extends StddevAttributeAggregator {
-        private final Attribute.Type type = Attribute.Type.INT;
+        private final Attribute.Type type = Attribute.Type.DOUBLE;
         private double mean, oldMean, stdDeviation, sum;
         private int count = 0;
 
         @Override
-        public Attribute.Type getReturnType() { return type; }
+        public Attribute.Type getReturnType() {
+            return type;
+        }
 
         @Override
         public Object processAdd(Object data) {
@@ -258,7 +284,7 @@ public class StddevAttributeAggregator extends AttributeAggregator {
                 oldMean = mean;
                 sum += value;
                 mean = sum / count;
-                stdDeviation += (value - oldMean)*(value - mean);
+                stdDeviation += (value - oldMean) * (value - mean);
             }
 
             if (count < 2) {
@@ -273,13 +299,13 @@ public class StddevAttributeAggregator extends AttributeAggregator {
             double value = (Integer) data;
 
             if (count == 0) {
-                sum = mean = 0;
-                stdDeviation = 0;
+                sum = mean = 0.0;
+                stdDeviation = 0.0;
             } else {
                 oldMean = mean;
                 sum -= value;
                 mean = sum / count;
-                stdDeviation -= (value - oldMean)*(value - mean);
+                stdDeviation -= (value - oldMean) * (value - mean);
             }
 
             if (count < 2) {
@@ -293,11 +319,13 @@ public class StddevAttributeAggregator extends AttributeAggregator {
             sum = mean = oldMean = 0.0;
             stdDeviation = 0.0;
             count = 0;
-            return 0;
+            return 0.0;
         }
 
         @Override
-        public Object[] currentState() { return new Object[] {sum, mean, oldMean, stdDeviation, count}; }
+        public Object[] currentState() {
+            return new Object[]{sum, mean, oldMean, stdDeviation, count};
+        }
 
         @Override
         public void restoreState(Object[] state) {
@@ -310,12 +338,14 @@ public class StddevAttributeAggregator extends AttributeAggregator {
     }
 
     private class StddevAttributeAggregatorLong extends StddevAttributeAggregator {
-        private final Attribute.Type type = Attribute.Type.LONG;
+        private final Attribute.Type type = Attribute.Type.DOUBLE;
         private double mean, oldMean, stdDeviation, sum;
         private int count = 0;
 
         @Override
-        public Attribute.Type getReturnType() { return type; }
+        public Attribute.Type getReturnType() {
+            return type;
+        }
 
         @Override
         public Object processAdd(Object data) {
@@ -330,7 +360,7 @@ public class StddevAttributeAggregator extends AttributeAggregator {
                 oldMean = mean;
                 sum += value;
                 mean = sum / count;
-                stdDeviation += (value - oldMean)*(value - mean);
+                stdDeviation += (value - oldMean) * (value - mean);
             }
 
             if (count < 2) {
@@ -345,13 +375,13 @@ public class StddevAttributeAggregator extends AttributeAggregator {
             double value = (Long) data;
 
             if (count == 0) {
-                sum = mean = 0;
-                stdDeviation = 0;
+                sum = mean = 0.0;
+                stdDeviation = 0.0;
             } else {
                 oldMean = mean;
                 sum -= value;
                 mean = sum / count;
-                stdDeviation -= (value - oldMean)*(value - mean);
+                stdDeviation -= (value - oldMean) * (value - mean);
             }
 
             if (count < 2) {
@@ -365,11 +395,13 @@ public class StddevAttributeAggregator extends AttributeAggregator {
             sum = mean = oldMean = 0.0;
             stdDeviation = 0.0;
             count = 0;
-            return 0;
+            return 0.0;
         }
 
         @Override
-        public Object[] currentState() { return new Object[] {sum, mean, oldMean, stdDeviation, count}; }
+        public Object[] currentState() {
+            return new Object[]{sum, mean, oldMean, stdDeviation, count};
+        }
 
         @Override
         public void restoreState(Object[] state) {
